==> scripts/02_Step_Distribution.Rmd <==

# 01.Load Packages
```{r, warning=FALSE}
library(readr)
library(haven)
library(dplyr)
library(ggplot2)
library(DataExplorer)
library(patchwork)
library(tidyr)
library(robustbase) # for Adaptive Trimmed Mean Estimator
library(moments)
library(nortest)
library(purrr)
library(broom)
library(blandr)
library(here)

source('01_Functions_Base.R')

custom_theme <- theme(
  panel.background = element_blank(),
  axis.ticks = element_blank(),
  axis.text = element_text(size = 10),
  axis.title = element_text(size = 12, face = "bold"),
  axis.line = element_line(size = 0.5),
  plot.title = element_text(size = 14, face = "bold", hjust = 0.5)
)

custom_theme2 <- theme(
    panel.grid.major = element_blank(),  
    panel.grid.minor = element_blank(),  
    panel.background = element_blank(),  
    axis.ticks = element_blank(),  
    panel.border = element_rect(colour = "black", fill = NA, size = 1)
  )

step_algorithms = c('ADEPT', 'Oak', 'SDT', 'Stepcount', 'Verisense')
```

# Clean and impute data
```{r}
dat_vis = read_csv(here('00_Intermediate_output/09_Step_Cov_WithoutSmokingImpute.csv'))
dat_vis$...1 = NULL
```

## Visualize extreme data
```{r}
## visulize the percentile
tmp = dat_vis[,c(1,3:7)] |>
  pivot_longer(cols = c(ADEPT, Oak, SDT, Stepcount, Verisense),
               names_to = "Algorithm",
               values_to = "Steps") |>
  group_by(Algorithm) |>
  summarise(`0-5000` = sum(Steps<=5000),
            `5001-10000` = sum(Steps <=10000 & Steps>5000),
            `10001-15000` = sum(Steps <=15000 & Steps>10000),
            `15001-20000` = sum(Steps <=20000 & Steps>15000),
            `20001-30000` = sum(Steps <=30000 & Steps>20000),
            `30001-40000` = sum(Steps <=40000 & Steps>30000),
            `40001-50000` = sum(Steps <=50000 & Steps>40000),
            `>50000` = sum(Steps >50000))

write_csv(tmp, here('Results/Tables/01_Step.distribution.beforeExtreme.csv'))
```

## Impute smoking statues
```{r}
## Impute missing smoking status.
dat_temp = cbind(dat_vis[, -which(colnames(dat_vis) == 'Smoking')], 
                 dat_vis[, which(colnames(dat_vis) == 'Smoking')])
colnames(dat_temp)[dim(dat_temp)[2]] = 'Smoking'
dat_temp$Smoking = as.factor(dat_temp$Smoking)
colnames(dat_temp)[1:(ncol(dat_temp)-1)] = paste0('var',c(1:(ncol(dat_temp)-1)))

library(mice)
dat_imp = mice(dat_temp,   # Include missing matrix
            m = 10,  # Imputation Number
            method = 'rf', # Imputation Methods
            printFlag = FALSE, # Not show history records
            seed = 208)

num = which(colnames(dat_temp) == 'Smoking')
dat_imp.all = do.call(cbind, 
                      lapply(1:10, function(i) complete(dat_imp, action = i)[, num, drop = FALSE]))

get_mode <- function(x) {
  uniqx <- unique(x)
  uniqx[which.max(tabulate(match(x, uniqx)))]
}

dat_temp$Smoking = apply(dat_imp.all[, 1:10], 1, get_mode)
dat_vis$Smoking = dat_temp$Smoking
write_csv(dat_vis, here('00_Intermediate_output/03_Step_Cov_Imputed.csv'))
```

## Exclude extreme values
by Adaptive Trimmed Mean Estimator
```{r}
result = read_csv(here('00_Intermediate_output/03_Step_Cov_Imputed.csv'))
# Identify extreme values for each step algorithm
ex.list <- unlist(lapply(step_algorithms, function(var) {
  adjbox_result <- adjbox(result[[var]], id.n = Inf, plot = TRUE)
  which(result[[var]] %in% adjbox_result$out)
}))

# Extract subject IDs with extreme values
ex.list <- unique(result$subjectid[ex.list])

# Filter out individuals with extreme values
result_no_extreme <- result %>%
  filter(!subjectid %in% ex.list)  # N = 1215; N = 37 Excluded

# Output the number of excluded individuals
cat('The number of individuals who are excluded due to extreme values: ',
    nrow(result) - nrow(result_no_extreme))  # Num = 30

# Save the dataset excluding extreme values to a CSV file
write_csv(result_no_extreme, 
          here('00_Intermediate_output/03_Step_Cov_Imputed_ExcludeExtreme.csv'))

# Reshape data for step count distribution analysis
result = read.csv(here('00_Intermediate_output/03_Step_Cov_Imputed_ExcludeExtreme.csv'))
result %>%
  select(subjectid, ADEPT, Oak, SDT, Stepcount, Verisense) %>%
  pivot_longer(cols = c(ADEPT, Oak, SDT, Stepcount, Verisense),
               names_to = "Algorithm",
               values_to = "Steps") %>%
  group_by(Algorithm) %>%
  summarise(
    `0-5000` = sum(Steps <= 5000),
    `5001-10000` = sum(Steps > 5000 & Steps <= 10000),
    `10001-15000` = sum(Steps > 10000 & Steps <= 15000),
    `15001-20000` = sum(Steps > 15000 & Steps <= 20000),
    `20001-30000` = sum(Steps > 20000 & Steps <= 30000),
    `30001-40000` = sum(Steps > 30000 & Steps <= 40000),
    `40001-50000` = sum(Steps > 40000 & Steps <= 50000),
    `>50000` = sum(Steps > 50000)
  )

write_csv(tmp, here('Results/Tables/01_Step.distribution.afterExtreme.csv'))
```

## Extract steps for distribution
```{r}
dat.step = read_csv(here('00_Intermediate_output/03_Step_Cov_Imputed.csv')) |>
  select(subjectid, ADEPT, Oak, SDT, Stepcount, Verisense)
```

### Skewness test
```{r}
variables_to_test <- step_algorithms
# Calculate skewness, kurtosis and conduct normality test.
skewness_results <- dat.step %>%
  select(all_of(variables_to_test)) %>%
  summarise(skewness = across(everything(), skewness))

kurtosis_results <- dat.step %>%
  select(all_of(variables_to_test)) %>%
  summarise(kurtosis = across(everything(), kurtosis))

shapiro_results <- dat.step %>%
  select(all_of(variables_to_test)) %>%
  map(shapiro.test) %>%
  map_dfr(tidy, .id = "Variable") %>%
  select(-statistic, method) %>%
  pivot_wider(names_from = Variable,
              values_from = p.value)

median_results <- dat.step |>
  select(all_of(variables_to_test)) %>%
  summarise(median = across(everything(), c(median)),
            IQR = across(everything(), c(IQR)))

result.temp = rbind(skewness_results$skewness, 
                    kurtosis_results$kurtosis, 
                    shapiro_results[,-1])
result.temp$Method = c('Skewness', 'Kurtosis', 'Shapiro-Wilk normality test')

writexl::write_xlsx(result.temp, here('Results/Tables/01_Distribution.xlsx'))
```


# 02. Scatter Plots and Correlation
## Combine Bland-Altman & Pairs Distributions 
### include Subtitle
Consider the space of visualization, I divide the number of steps by 1,000
```{r}
df = dat.step[,-1]/1000

# Row:i, Column:j
custom_theme <- theme_minimal(base_size = 15) + 
  theme(
    plot.title = element_text(hjust = 0.5, face = "bold", size = 10),
    axis.text = element_text(size = 9),
    axis.title = element_text(size = 9),
    panel.grid.major = element_line(color = "grey90"),
    panel.border = element_rect(color = "black", fill = NA),
    #panel.grid.major = element_blank(),  
    panel.grid.minor = element_blank(),  
    axis.ticks = element_blank(),
    plot.subtitle = element_text(size = 8, face = "italic", hjust = 0.5),
    plot.margin = margin(5, 5, 5, 5))

# Initializes the drawing list
plot.all <- list()

# Iterate through all step_algorithms, generating a graph for each location
for (i in 1:length(step_algorithms)) {
  for (j in 1:length(step_algorithms)) {
    
    # Diagonal i = j: histogram
    if (i == j) {
      plot.all[[paste0('H', i, i)]] <- ggplot(df, aes(x = !!sym(step_algorithms[i]))) +
        geom_histogram(
          aes(y = ..count..), 
          bins = 15, 
          fill = "#3D72BE", 
          color = "white"
        ) + labs(
          title = paste0(step_algorithms[i]),
          x = "Step Counts",
          y = "Frequency",
          subtitle = "Histogram"
        ) + custom_theme
      next
    }
    
    # Upper triangular: Draw the Bland-Altman plot
    if (i < j) {
      # Calculate Bland-Altman data
      ba_data <- blandr.statistics(df[[step_algorithms[i]]], 
                                   df[[step_algorithms[j]]])
      
      # Dataframe with differences and averages
      plot_data <- data.frame(
        mean = (df[[step_algorithms[i]]] + df[[step_algorithms[j]]]) / 2,
        difference = df[[step_algorithms[i]]] - df[[step_algorithms[j]]]
      )
      # Linear regression
      lm_fit <- lm(difference ~ mean, data = plot_data)
      lm_coef <- tidy(lm_fit)
      
      # Extract Coefficients
      intercept <- lm_coef$estimate[1]
      slope <- lm_coef$estimate[2]
      
      # Create Bland-Altman; size = 0.7
      p <- ggplot(plot_data, aes(x = mean, y = difference)) +
        geom_point(alpha = 0.3, color = 'black') +
        geom_hline(yintercept = ba_data$bias, color = "blue", linetype = "dashed") +
        geom_hline(yintercept = ba_data$upperLOA, color = "red", linetype = "dashed") +
        geom_hline(yintercept = ba_data$lowerLOA, color = "red", linetype = "dashed") +
        geom_smooth(method = "gam", formula = y ~ s(x, bs = "cs", k = 5),
                    color = "#E74C3C",
                    se = TRUE, span = 2, fill = '#E74C3C') +  
        labs(
          x = paste0("Mean of ", step_algorithms[i], " and ", step_algorithms[j]), 
          y = paste0("Difference (", step_algorithms[i], " - ", step_algorithms[j], ")"),
          title = paste0(step_algorithms[i], " ~ ", step_algorithms[j]),
          subtitle = "Bland-Altman plot"
        ) + custom_theme
      
      plot.all[[paste0('BA', i, j)]] <- p
      next
    }
    
    # Lower triangular position i>j: pairwise scatter plots
    if (i > j) {
      correlation <- cor(df[[step_algorithms[i]]], df[[step_algorithms[j]]])
      
      plot.all[[paste0('PP', i, j)]] <- ggplot(df, aes(
        x = !!sym(step_algorithms[i]), y = !!sym(step_algorithms[j])
      )) +
        geom_point(color = "#E74C3C", alpha = 0.3) +
        geom_smooth(
          method = "gam", formula = y ~ s(x, bs = "cs", k = 5), 
          color = "black", size = 1, se = TRUE, fill = "black"
        ) +
        custom_theme + 
        labs(
          title = paste0(step_algorithms[i], " ~ ", step_algorithms[j]),
          x = step_algorithms[i],
          y = step_algorithms[j],
          subtitle = "Pairwise Distributions"
        ) + 
        annotate("text", 
                 x = min(df[[step_algorithms[i]]]), 
                 y = Inf, 
                 label = paste0("r = ", round(correlation, 2)), 
           hjust = 0, vjust = 2.0, size = 4, color = "black", fontface = "bold")
        # scale_y_continuous(limits = c(min(df[[step_algorithms[i]]], df[[step_algorithms[j]]]),
        #                               max(df[[step_algorithms[i]]], df[[step_algorithms[j]]]))) + 
        # scale_x_continuous(limits = c(min(df[[step_algorithms[i]]], df[[step_algorithms[j]]]),
        #                               max(df[[step_algorithms[i]]], df[[step_algorithms[j]]])))
      next
    }
  }
}

combined_plot <- wrap_plots(
  plot.all$H11, plot.all$BA12, plot.all$BA13, plot.all$BA14, plot.all$BA15,
  plot.all$PP21, plot.all$H22, plot.all$BA23, plot.all$BA24, plot.all$BA25,
  plot.all$PP31, plot.all$PP32, plot.all$H33, plot.all$BA34, plot.all$BA35,
  plot.all$PP41, plot.all$PP42, plot.all$PP43, plot.all$H44, plot.all$BA45,
  plot.all$PP51, plot.all$PP52, plot.all$PP53, plot.all$PP54, plot.all$H55,
  ncol = 5
) & 
theme(legend.position = 'bottom')

png(here('Results/Figures/03_Combined.subtitle.png'),
    width = 12, height = 14, units = 'in', res = 300)
print(combined_plot)
dev.off()
```


### No Subtitle
Consider the space of visualization, I divide the number of steps by 1,000
```{r}
df = dat.step[,-1]/1000

# Row:i, Column:j
custom_theme <- theme_minimal(base_size = 15) + 
  theme(
    plot.title = element_text(hjust = 0.5, face = "bold", size = 10),
    axis.text = element_text(size = 9),
    axis.title = element_text(size = 9),
    panel.grid.major = element_line(color = "grey90"),
    panel.border = element_rect(color = "black", fill = NA),
    #panel.grid.major = element_blank(),  
    panel.grid.minor = element_blank(),  
    axis.ticks = element_blank(),
    plot.subtitle = element_text(size = 8, face = "italic", hjust = 0.5),
    plot.margin = margin(6, 6, 6, 6))

# Initializes the drawing list
plot.all <- list()

# Iterate through all step_algorithms, generating a graph for each location
for (i in 1:length(step_algorithms)) {
  for (j in 1:length(step_algorithms)) {
    
    # Diagonal i = j: histogram
    if (i == j) {
      plot.all[[paste0('H', i, i)]] <- ggplot(df, aes(x = !!sym(step_algorithms[i]))) +
        geom_histogram(
          aes(y = ..count..), 
          bins = 15, 
          fill = "#3D72BE",  #skyblue
          color = "white"
        ) + labs(
          title = paste0(step_algorithms[i]),
          x = "Step Counts",
          y = "Frequency"
        ) + custom_theme
      next
    }
    
    # Upper triangular position i<j: Bland-Altman
    if (i < j) {
      # Calculate Bland-Altman data
      ba_data <- blandr.statistics(df[[step_algorithms[i]]], 
                                   df[[step_algorithms[j]]])
      
      # Dataframe with differences and averages
      plot_data <- data.frame(
        mean = (df[[step_algorithms[i]]] + df[[step_algorithms[j]]]) / 2,
        difference = df[[step_algorithms[i]]] - df[[step_algorithms[j]]]
      )
      # Linear regression
      lm_fit <- lm(difference ~ mean, data = plot_data)
      lm_coef <- tidy(lm_fit)
      
      # Extract Coefficients
      intercept <- lm_coef$estimate[1]
      slope <- lm_coef$estimate[2]
      
      # Create Bland-Altman
      p <- ggplot(plot_data, aes(x = mean, y = difference)) +
        geom_point(alpha = 0.3, color = 'black') +
        geom_hline(yintercept = ba_data$bias, color = "blue", linetype = "dashed") +
        geom_hline(yintercept = ba_data$upperLOA, color = "red", linetype = "dashed") +
        geom_hline(yintercept = ba_data$lowerLOA, color = "red", linetype = "dashed") +
        geom_smooth(method = "loess", color = "#E74C3C", se = TRUE, span = 2, fill = '#E74C3C') +  
        labs(
          x = paste0("Mean of ", step_algorithms[i], " and ", step_algorithms[j]), 
          y = paste0("Difference (", step_algorithms[i], " - ", step_algorithms[j], ")"),
          title = paste0(step_algorithms[i], " ~ ", step_algorithms[j])
        ) + custom_theme
      
      plot.all[[paste0('BA', i, j)]] <- p
      next
    }
    
    # Lower triangular position i>j: pairwise scatter plots
    if (i > j) {
      correlation <- cor(df[[step_algorithms[i]]], df[[step_algorithms[j]]])
      
      plot.all[[paste0('PP', i, j)]] <- ggplot(df, aes(
        x = !!sym(step_algorithms[i]), y = !!sym(step_algorithms[j])
      )) +
        geom_point(color = "#E74C3C", alpha = 0.3) +
        geom_smooth(
          method = "gam", formula = y ~ s(x, bs = "cs", k = 5), 
          color = "black", size = 1, se = TRUE, fill = "black"
        ) +
        custom_theme + 
        labs(
          title = paste0(step_algorithms[i], " ~ ", step_algorithms[j]),
          x = step_algorithms[i],
          y = step_algorithms[j]
        ) + annotate("text", 
                 x = min(df[[step_algorithms[i]]]), 
                 y = Inf, 
                 label = paste0("r = ", round(correlation, 2)), 
           hjust = 0, vjust = 2.0, size = 4, color = "black", fontface = "bold")
        # scale_y_continuous(limits = c(min(df[[step_algorithms[i]]], df[[step_algorithms[j]]]),
        #                               max(df[[step_algorithms[i]]], df[[step_algorithms[j]]]))) + 
        # scale_x_continuous(limits = c(min(df[[step_algorithms[i]]], df[[step_algorithms[j]]]),
        #                               max(df[[step_algorithms[i]]], df[[step_algorithms[j]]])))
      next
    }
  }
}

combined_plot <- wrap_plots(
  plot.all$H11, plot.all$BA12, plot.all$BA13, plot.all$BA14, plot.all$BA15,
  plot.all$PP21, plot.all$H22, plot.all$BA23, plot.all$BA24, plot.all$BA25,
  plot.all$PP31, plot.all$PP32, plot.all$H33, plot.all$BA34, plot.all$BA35,
  plot.all$PP41, plot.all$PP42, plot.all$PP43, plot.all$H44, plot.all$BA45,
  plot.all$PP51, plot.all$PP52, plot.all$PP53, plot.all$PP54, plot.all$H55,
  ncol = 5
) & 
theme(legend.position = 'bottom')

png(here('Results/Figures/03_Combined.png'),
    width = 12, height = 12, units = 'in', res = 300)
print(combined_plot)
dev.off()
```

### No Subtitle + Scaled
Consider the space of visualization, I divide the number of steps by 1,000
```{r}
df = as.data.frame(scale(dat.step[,-1]))

# Row:i, Column:j
custom_theme <- theme_minimal(base_size = 15) + 
  theme(
    plot.title = element_text(hjust = 0.5, face = "bold", size = 10),
    axis.text = element_text(size = 9),
    axis.title = element_text(size = 9),
    panel.grid.major = element_line(color = "grey90"),
    panel.border = element_rect(color = "black", fill = NA),
    #panel.grid.major = element_blank(),  
    panel.grid.minor = element_blank(),  
    axis.ticks = element_blank(),
    plot.subtitle = element_text(size = 8, face = "italic", hjust = 0.5),
    plot.margin = margin(6, 6, 6, 6))

# Initializes the drawing list
plot.all <- list()

# Iterate through all step_algorithms, generating a graph for each location
for (i in 1:length(step_algorithms)) {
  for (j in 1:length(step_algorithms)) {
    
    # Diagonal i = j: histogram
    if (i == j) {
      plot.all[[paste0('H', i, i)]] <- ggplot(df, aes(x = !!sym(step_algorithms[i]))) +
        geom_histogram(
          aes(y = ..count..), 
          bins = 15, 
          fill = "#3D72BE",  #skyblue 
          color = "white"
        ) + labs(
          title = paste0(step_algorithms[i]),
          x = "Step Counts",
          y = "Frequency"
        ) + custom_theme
      next
    }
    
    # Upper triangular position i<j: Bland-Altman
    if (i < j) {
      # Calculate Bland-Altman data
      ba_data <- blandr.statistics(df[,i], 
                                   df[,j])
      
      # Dataframe with differences and averages
      plot_data <- data.frame(
        mean = (df[,i] + df[,j]) / 2,
        difference = df[,i] - df[,j]
      )
      # Linear regression
      lm_fit <- lm(difference ~ mean, data = plot_data)
      lm_coef <- tidy(lm_fit)
      
      # Extract Coefficients
      intercept <- lm_coef$estimate[1]
      slope <- lm_coef$estimate[2]
      
      # Create Bland-Altman # #3D72BE, #ED7720
      p <- ggplot(plot_data, aes(x = mean, y = difference)) +
        geom_point(alpha = 0.3, color = 'black') +
        geom_hline(yintercept = ba_data$bias, color = "blue", linetype = "dashed") +
        geom_hline(yintercept = ba_data$upperLOA, color = "red", linetype = "dashed") +
        geom_hline(yintercept = ba_data$lowerLOA, color = "red", linetype = "dashed") +
        geom_smooth(method = "loess", color = "#E74C3C", se = TRUE, span = 2, fill='#E74C3C') +  
        labs(
          x = paste0("Mean of ", step_algorithms[i], " and ", step_algorithms[j]), 
          y = paste0("Difference (", step_algorithms[i], " - ", step_algorithms[j], ")"),
          title = paste0(step_algorithms[i], " ~ ", step_algorithms[j])
        ) + custom_theme
      
      plot.all[[paste0('BA', i, j)]] <- p
      next
    }
    
    # Lower triangular position i>j: pairwise scatter plots
    if (i > j) {
      correlation <- cor(df[,i], df[,j])
      
      plot.all[[paste0('PP', i, j)]] <- ggplot(df, aes(
        x = !!sym(step_algorithms[i]), y = !!sym(step_algorithms[j])
      )) +
        geom_point(color = "#E74C3C", alpha = 0.3) +
        geom_smooth(
          method = "gam", formula = y ~ s(x, bs = "cs", k = 5), 
          color = "black", size = 1, se = TRUE, fill = "black"
        ) +
        custom_theme + 
        labs(
          title = paste0(step_algorithms[i], " ~ ", step_algorithms[j]),
          x = step_algorithms[i],
          y = step_algorithms[j]
        ) + 
        annotate("text", 
                 x = min(df[,i]), 
                 y = Inf, 
                 label = paste0("r = ", round(correlation, 2)), 
           hjust = 0, vjust = 2.0, size = 4, color = "black", fontface = "bold") + 
        scale_y_continuous(limits = c(min(df[[step_algorithms[i]]], df[[step_algorithms[j]]]),
                                      max(df[[step_algorithms[i]]], df[[step_algorithms[j]]]))) + 
        scale_x_continuous(limits = c(min(df[[step_algorithms[i]]], df[[step_algorithms[j]]]),
                                      max(df[[step_algorithms[i]]], df[[step_algorithms[j]]])))
      next
    }
  }
}

combined_plot <- wrap_plots(
  plot.all$H11, plot.all$BA12, plot.all$BA13, plot.all$BA14, plot.all$BA15,
  plot.all$PP21, plot.all$H22, plot.all$BA23, plot.all$BA24, plot.all$BA25,
  plot.all$PP31, plot.all$PP32, plot.all$H33, plot.all$BA34, plot.all$BA35,
  plot.all$PP41, plot.all$PP42, plot.all$PP43, plot.all$H44, plot.all$BA45,
  plot.all$PP51, plot.all$PP52, plot.all$PP53, plot.all$PP54, plot.all$H55,
  ncol = 5
) & 
theme(legend.position = 'bottom')

png(here('Results/Figures/03_Combined(Scaled).png'), width = 12, height = 12, units = 'in', res = 300)
print(combined_plot)
dev.off()
```

# 03.Distribution ~ demographical variables 
```{r}
dat_vis = read_csv(here('00_Intermediate_output/03_Step_Cov_Imputed.csv'))
```

## Single Covariate
### Continous variables
Ask Siyu for help (Smooth them).
```{r}
# Define continuous covariates for plotting
cov_continuous <- c("Age") # Add other continuous covariates like "SPPB" if needed
plot_list_all <- list()

# Generate plots for each covariate and step algorithm
for (covariate in cov_continuous) {
  plot_list <- lapply(step_algorithms, function(var) {
    ggplot(dat_vis, aes(x = .data[[covariate]], y = .data[[var]])) +
      geom_smooth(se = TRUE, color = "#3D72BE") + # Smoothed trend line
      labs(
        x = ifelse(covariate == "Age", "Age (years)", "SPPB (score)"),
        y = "Step Counts",
        title = var
      ) +
      theme_minimal() + 
      custom_theme +
      # scale_y_continuous(limits = case_when(var=='ADEPT'~c(0,2000),
      #                                       var=='Oak'~c(1000,8000),
      #                                       var=='SDT'~c(15000,30000),
      #                                       var=='Stepcount'~c(1000,7000),
      #                                       var=='Verisense'~c(1000,5000)),
      #                    breaks = function(x) pretty(x, n = 3))+
      scale_y_continuous(breaks = pretty) +
      scale_x_continuous(breaks = pretty)
  })
  
  names(plot_list) <- paste0(step_algorithms, '-', covariate)
  plot_list_all <- c(plot_list_all, plot_list)
}

# Combine all plots into one grid
combined_plot <- wrap_plots(plot_list_all, ncol = 5) + 
  plot_layout(guides = "collect") & 
  theme(legend.position = 'bottom')

# Save combined plot to file
png(here('Results/Figures/04_Continuous_Steps.png'), height = 5, width = 10, units = 'in', res = 300)
print(combined_plot)
dev.off()
```

## Associations with Age group by covariates
```{r}
"#8E44AD"
"#4CAF50"
"#F4D03F"
"#E74C3C"
"#95A5A6"
"#97470C" 
"#ED7720"
"#D81B60"
"#1E88E5"
"#FFC107"


colors <- c("#ED7720", "#3D72BE", "#E74C3C","#97470C","#4CAF50", "#F4D03F", "#95A5A6")
```

### Sex, Race, Education, Smoking, Drinking
Latest edition: change from loess to linear.
```{r}
calculate_combined_slope <- function(model, coefficients) {
  beta <- coef(model)[coefficients] # Extract coefficient estimates
  combined_slope <- sum(beta) # Calculate combined slope
  
  # Extract variance-covariance matrix
  vcov_matrix <- vcov(model)[coefficients, coefficients]
  
  # Calculate standard error of combined slope
  se_combined <- sqrt(sum(vcov_matrix))
  
  # Calculate confidence interval
  t_value <- qt(0.975, df = df.residual(model))
  ci_lower <- combined_slope - t_value * se_combined
  ci_upper <- combined_slope + t_value * se_combined
  
  # Calculate t-statistic
  t_statistic <- combined_slope / se_combined
  df <- df.residual(model)
  
  # Calculate p-value (two-tailed test)
  p_value <- 2 * pt(-abs(t_statistic), df)
  
  # Return results
  return(list(
    slope = combined_slope,
    se = se_combined,
    t_statistic = t_statistic,
    p_value = p_value,
    ci_lower = ci_lower,
    ci_upper = ci_upper,
    df = df
  ))
}

# Adjust the position of factor variable
dat_vis$Education = factor(dat_vis$Education,
                           levels = c('Basic Education', 
                                      'Intermediate Education', 
                                      'Advanced Education'))

dat_vis$APOE = factor(dat_vis$APOE,
                      levels = c('No ε4 allele', 
                                      '≥1 APOE ε4 allele'))

factor_columns <- c('Sex', 'Race', 'Smoking', 'Drinking', 'APOE', 'Education')
dat_vis[factor_columns] <- lapply(dat_vis[factor_columns], as.factor)
```

### Generate Linear Coefficients for each lines
#### Non-scaling
```{r, warning=FALSE}
cov_continous <- c("Sex", 'Race', 'Education', "Smoking", "Drinking", 'APOE')
label = list('exam_age:SexMale', 
             'exam_age:RaceWhite',
             c('exam_age:EducationIntermediate Education', 'exam_age:EducationAdvanced Education'),
             c('exam_age:SmokingFormer smoker', 'exam_age:SmokingNever smoker'),
             c('exam_age:DrinkingFormer drinker','exam_age:DrinkingNever drinker'),
             'exam_age:APOE≥1 APOE ε4 allele')
          
fit_results <- data.frame()
for (covariate in cov_continous) {
  plot_list = list()
  tmp_summary <- data.frame()
  
  for (var in step_algorithms){
    temp = dat_vis[!is.na(dat_vis[[covariate]]), ]
    name = step_algorithms[which(step_algorithms == var)]
    plot = ggplot(temp, aes(x = exam_age, y = .data[[var]], 
                            color = as.factor(get(covariate)),
                            fill = as.factor(get(covariate)))) +
      geom_smooth(method = "lm", se = TRUE, alpha = 0.2) +
      scale_color_manual(values = colors) +
      scale_fill_manual(values = colors) +
      labs(x = 'Age (year)', y = 'Step Counts', #paste0("Step counts (", name, ')'
           fill = covariate,
           color = covariate,
           title = name) +
      theme_minimal() + 
      custom_theme
    
    plot_list[[var]] = plot
    
    # Fit linear model and extract results
    fit <- lm(formula = as.formula(paste(var, "~ exam_age * ", covariate)), data = temp)
    fit_summary <- tidy(fit)
    fit_summary$ci_lower = confint(fit)[,1]
    fit_summary$ci_upper = confint(fit)[,2]
    fit_summary$method = step_algorithms[which(var == step_algorithms)]
    
    num = which(covariate == cov_continous)
    tmp_summary = rbind(tmp_summary, fit_summary[2,])
    for (i in 1:length(label[[num]])){
      tmp <- calculate_combined_slope(fit, c('exam_age', 
                                             label[[num]][i]))
      tmp_summary <- rbind(tmp_summary,
                           c(label[[num]][i], tmp$slope, tmp$se, tmp$t_statistic,
                             tmp$p_value, tmp$ci_lower, tmp$ci_upper, 
                             step_algorithms[which(var == var_list)]))
    }
    colnames(tmp_summary) = colnames(fit_summary)
  }
  ## Plots
  output_name = paste0('Results/Figures/05_Age_GroupBy_', covariate, '_LR.png')
  png(here(output_name), height = 4, width = 10, units = 'in', res = 300)
  a = wrap_plots(plot_list, ncol = 5) + 
    plot_layout(guides = "collect") &
    theme(legend.position = 'bottom',
          legend.text = element_text(size = 9),
          legend.title = element_text(size = 9) )
  print(a)
  dev.off()
  
  ## Modify Results
  tmp_summary[, 2:7] = sapply(tmp_summary[, 2:7], as.numeric)
  tmp_summary = tmp_summary |> 
    mutate(`Conf (95%CI)` = sprintf("%.1f (%.1f, %.1f)", estimate, ci_lower, ci_upper),
           Pvalue = sprintf("%.3f", p.value)) |>
    select(term, `Conf (95%CI)`, 'method', 'Pvalue')
  
  fit_results <- rbind(fit_results, tmp_summary)
}

writexl::write_xlsx(fit_results, here('Results/Tables/05_GroupBy_Conefficients.xlsx'))
```

#### Scaling
```{r, warning=FALSE}
dat.vis.scaling = dat_vis
dat.vis.scaling[,3:7] = scale(dat.vis.scaling[,3:7])

fit_results <- data.frame()
for (covariate in cov_continous) {
  plot_list = list()
  tmp_summary <- data.frame()
  
  for (var in step_algorithms){
    temp = dat.vis.scaling[!is.na(dat.vis.scaling[[covariate]]), ]
    name = step_algorithms[which(step_algorithms == var)]
    plot = ggplot(temp, aes(x = exam_age, y = .data[[var]], 
                            color = as.factor(get(covariate)),
                            fill = as.factor(get(covariate)))) +
      geom_smooth(method = "lm", se = TRUE, alpha = 0.2) +
      scale_color_manual(values = colors) +
      scale_fill_manual(values = colors) +
      labs(x = 'Age (year)', y = 'Step Counts', #paste0("Step counts (", name, ')'
           fill = covariate,
           color = covariate,
           title = name) +
      theme_minimal() + 
      custom_theme
    
    plot_list[[var]] = plot
    
    # Fit linear model and extract results
    fit <- lm(formula = as.formula(paste(var, "~ exam_age * ", covariate)), data = temp)
    fit_summary <- tidy(fit)
    fit_summary$ci_lower = confint(fit)[,1]
    fit_summary$ci_upper = confint(fit)[,2]
    fit_summary$method = step_algorithms[which(var == step_algorithms)]
    
    num = which(covariate == cov_continous)
    tmp_summary = rbind(tmp_summary, fit_summary[2,])
    for (i in 1:length(label[[num]])){
      tmp <- calculate_combined_slope(fit, c('exam_age', 
                                             label[[num]][i]))
      tmp_summary <- rbind(tmp_summary,
                           c(label[[num]][i], tmp$slope, tmp$se, tmp$t_statistic,
                             tmp$p_value, tmp$ci_lower, tmp$ci_upper, 
                             step_algorithms[which(var == var_list)]))
    }
    colnames(tmp_summary) = colnames(fit_summary)
  }
  ## Plots
  output_name = paste0('Results/Figures/05_Age_GroupBy_Covariates/05_Age_GroupBy_', 
                       covariate, '_LR(scaling).png')
  png(here(output_name), height = 4, width = 10, units = 'in', res = 300)
  a = wrap_plots(plot_list, ncol = 5) + 
    plot_layout(guides = "collect") &
    theme(legend.position = 'bottom',
          legend.text = element_text(size = 9),
          legend.title = element_text(size = 9) )
  print(a)
  dev.off()
  
  ## Modify Results
  tmp_summary[, 2:7] = sapply(tmp_summary[, 2:7], as.numeric)
  tmp_summary = tmp_summary |> 
    mutate(`Conf (95%CI)` = sprintf("%.2f (%.2f, %.2f)", estimate, ci_lower, ci_upper),
           Pvalue = sprintf("%.3f", p.value)) |>
    select(term, `Conf (95%CI)`, 'method', 'Pvalue')
  
  fit_results <- rbind(fit_results, tmp_summary)
}

writexl::write_xlsx(fit_results,
                    here('Results/Figures/05_Age_GroupBy_Covariates/05_GroupBy_Conefficients(Scaling).xlsx'))
```
